refactor: split statement EIR lowering
