+++ src/ir_lower/stmt/instance_property_writes.rs
@@ -1,0 +1,207 @@
+//! Purpose:
+//! Instance property writes, magic setters, and hook dispatch.
+//!
+//! Called from:
+//! - `crate::ir_lower::stmt`.
+//!
+//! Key details:
+//! - Preserves statement ordering, CFG shape, EIR effects, and ownership contracts.
+
+use super::*;
+
+/// Lowers an object property write.
+pub(super) fn lower_property_assign(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: &Expr,
+    property: &str,
+    value: &Expr,
+    span: Span,
+) {
+    // A statically-decided readonly-property write outside the declaring
+    // constructor raises a catchable `Error` in PHP rather than a compile-time
+    // error, but the object and RHS expressions must still be evaluated first.
+    let throw_access_message = ctx.throw_access_sites.get(&span).and_then(|info| {
+        if let ThrowAccessKind::ReadonlyProperty { class_name, property } = &info.kind {
+            Some(format!("Cannot modify readonly property {}::${}", class_name, property))
+        } else {
+            None
+        }
+    });
+    let object = lower_expr(ctx, object);
+    let value_expr = value;
+    let lowered_value = lower_expr(ctx, value_expr);
+    if let Some(message) = throw_access_message {
+        if ctx.value_is_owning_temporary(object) {
+            crate::ir_lower::ownership::release_if_owned(ctx, object, Some(span));
+        }
+        if ctx.value_is_owning_temporary(lowered_value) {
+            crate::ir_lower::ownership::release_if_owned(ctx, lowered_value, Some(span));
+        }
+        lower_throw_access_error(ctx, &message, span);
+        return;
+    }
+    let value = contextualize_property_array_assignment(
+        ctx,
+        object.value,
+        property,
+        lowered_value,
+        value_expr,
+        span,
+    );
+    if magic_set_receiver_has_method(ctx, object.value, property) {
+        lower_magic_property_set(ctx, object.value, property, value, span);
+        return;
+    }
+    // Route a write to a set-hooked property to its `__propset_<p>($value)` accessor, except inside
+    // that property's own accessor where `$this->prop = v` must write the raw backing slot.
+    if set_hook_receiver_has_accessor(ctx, object.value, property)
+        && !ctx.in_own_property_accessor(property)
+    {
+        lower_property_hook_set(ctx, object.value, property, value, span);
+        return;
+    }
+    let data = ctx.intern_string(property);
+    ctx.emit_void(
+        Op::PropSet,
+        vec![object.value, value.value],
+        Some(Immediate::Data(data)),
+        Op::PropSet.default_effects(),
+        Some(span),
+    );
+    if let Some(property_ty) = object_property_type(ctx, object.value, property) {
+        release_property_assignment_source_after_retaining_store(ctx, &property_ty, value, span);
+    }
+}
+
+/// Returns true when a property write should dispatch to `__set`.
+pub(super) fn magic_set_receiver_has_method(
+    ctx: &LoweringContext<'_, '_>,
+    object: crate::ir::ValueId,
+    property: &str,
+) -> bool {
+    let PhpType::Object(class_name) = ctx.builder.value_php_type(object).codegen_repr() else {
+        return false;
+    };
+    let normalized = class_name.trim_start_matches('\\');
+    let Some(class_info) = ctx.classes.get(normalized) else {
+        return false;
+    };
+    if class_info
+        .properties
+        .iter()
+        .any(|(name, _)| name == property)
+    {
+        return false;
+    }
+    class_info.methods.contains_key(&php_symbol_key("__set"))
+}
+
+/// Lowers an undeclared property write to a normal `__set` instance-method call.
+pub(super) fn lower_magic_property_set(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: crate::ir::ValueId,
+    property: &str,
+    value: LoweredValue,
+    span: Span,
+) {
+    let property_data = ctx.intern_string(property);
+    let property_name = ctx.emit_value(
+        Op::ConstStr,
+        Vec::new(),
+        Some(Immediate::Data(property_data)),
+        PhpType::Str,
+        Op::ConstStr.default_effects(),
+        Some(span),
+    );
+    let method_data = ctx.intern_string("__set");
+    ctx.emit_void(
+        Op::MethodCall,
+        vec![object, property_name.value, value.value],
+        Some(Immediate::Data(method_data)),
+        Op::MethodCall.default_effects(),
+        Some(span),
+    );
+    release_magic_set_value_after_call(ctx, value, span);
+}
+
+/// Releases an owning RHS temporary after the `__set` call has consumed it.
+pub(super) fn release_magic_set_value_after_call(
+    ctx: &mut LoweringContext<'_, '_>,
+    value: LoweredValue,
+    span: Span,
+) {
+    if ctx.value_is_owning_temporary(value) {
+        crate::ir_lower::ownership::release_if_owned(ctx, value, Some(span));
+    }
+}
+
+/// Returns true when the runtime class of `object` declares a `__propset_<property>` set-hook
+/// accessor, meaning a write to `property` should be routed through it.
+pub(super) fn set_hook_receiver_has_accessor(
+    ctx: &LoweringContext<'_, '_>,
+    object: crate::ir::ValueId,
+    property: &str,
+) -> bool {
+    let PhpType::Object(class_name) = ctx.builder.value_php_type(object).codegen_repr() else {
+        return false;
+    };
+    let normalized = class_name.trim_start_matches('\\');
+    ctx.classes.get(normalized).is_some_and(|info| {
+        info.methods
+            .contains_key(&php_symbol_key(&property_hook_set_method(property)))
+    })
+}
+
+/// Lowers a write to a set-hooked property as a call to its `__propset_<p>($value)` accessor,
+/// passing the assigned value as the single argument and releasing it if it was an owning temporary.
+pub(super) fn lower_property_hook_set(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: crate::ir::ValueId,
+    property: &str,
+    value: LoweredValue,
+    span: Span,
+) {
+    let method_data = ctx.intern_string(&property_hook_set_method(property));
+    ctx.emit_void(
+        Op::MethodCall,
+        vec![object, value.value],
+        Some(Immediate::Data(method_data)),
+        Op::MethodCall.default_effects(),
+        Some(span),
+    );
+    release_magic_set_value_after_call(ctx, value, span);
+}
+
+/// Converts array literals to hash storage when a declared object property requires assoc storage.
+pub(super) fn contextualize_property_array_assignment(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: crate::ir::ValueId,
+    property: &str,
+    lowered: LoweredValue,
+    value_expr: &Expr,
+    span: Span,
+) -> LoweredValue {
+    let php_type = ctx.builder.value_php_type(lowered.value);
+    if !matches!(value_expr.kind, ExprKind::ArrayLiteral(_)) {
+        return lowered;
+    }
+    if !matches!(php_type.codegen_repr(), PhpType::Array(_)) {
+        return lowered;
+    }
+    let Some(contextual_ty) = object_property_type(ctx, object, property) else {
+        return lowered;
+    };
+    let contextual_ty = contextual_ty.codegen_repr();
+    if !matches!(contextual_ty, PhpType::AssocArray { .. }) {
+        return lowered;
+    }
+    ctx.emit_value(
+        Op::ArrayToHash,
+        vec![lowered.value],
+        None,
+        contextual_ty,
+        Op::ArrayToHash.default_effects(),
+        Some(span),
+    )
+}
+
