+++ src/ir_lower/stmt/property_array_writes.rs
@@ -1,0 +1,270 @@
+//! Purpose:
+//! Instance property array mutations and retaining-store cleanup.
+//!
+//! Called from:
+//! - `crate::ir_lower::stmt`.
+//!
+//! Key details:
+//! - Preserves statement ordering, CFG shape, EIR effects, and ownership contracts.
+
+use super::*;
+
+/// Lowers `$object->prop[] = value`.
+pub(super) fn lower_property_array_push(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: &Expr,
+    property: &str,
+    value: &Expr,
+    span: Span,
+) {
+    let object = lower_expr(ctx, object);
+    if let Some(property_ty) =
+        object_property_type(ctx, object.value, property).filter(is_indexed_array_type)
+    {
+        let data = ctx.intern_string(property);
+        let property_value = ctx.emit_value(
+            Op::PropGet,
+            vec![object.value],
+            Some(Immediate::Data(data)),
+            property_ty.clone(),
+            Op::PropGet.default_effects(),
+            Some(span),
+        );
+        let property_value =
+            crate::ir_lower::ownership::acquire_if_refcounted(ctx, property_value, Some(span));
+        let value = lower_expr(ctx, value);
+        ctx.emit_void(
+            Op::ArrayPush,
+            vec![property_value.value, value.value],
+            None,
+            Op::ArrayPush.default_effects(),
+            Some(span),
+        );
+        release_property_array_insert_value_after_retain(ctx, &property_ty, value, span);
+        ctx.emit_void(
+            Op::PropSet,
+            vec![object.value, property_value.value],
+            Some(Immediate::Data(data)),
+            Op::PropSet.default_effects(),
+            Some(span),
+        );
+        release_rewritten_property_value_after_retaining_store(
+            ctx,
+            &property_ty,
+            property_value,
+            span,
+        );
+        return;
+    }
+
+    let value = lower_expr(ctx, value);
+    let data = ctx.intern_string(property);
+    ctx.emit_void(
+        Op::RuntimeCall,
+        vec![object.value, value.value],
+        Some(Immediate::Data(data)),
+        effects_lookup::runtime_effects(),
+        Some(span),
+    );
+}
+
+/// Lowers `$object->prop[index] = value`.
+pub(super) fn lower_property_array_assign(
+    ctx: &mut LoweringContext<'_, '_>,
+    object: &Expr,
+    property: &str,
+    index: &Expr,
+    value: &Expr,
+    span: Span,
+) {
+    let object = lower_expr(ctx, object);
+    if let Some(property_ty) =
+        object_property_type(ctx, object.value, property).filter(is_indexed_array_type)
+    {
+        let data = ctx.intern_string(property);
+        let property_value = ctx.emit_value(
+            Op::PropGet,
+            vec![object.value],
+            Some(Immediate::Data(data)),
+            property_ty.clone(),
+            Op::PropGet.default_effects(),
+            Some(span),
+        );
+        let property_value =
+            crate::ir_lower::ownership::acquire_if_refcounted(ctx, property_value, Some(span));
+        let index = lower_expr(ctx, index);
+        let value = lower_expr(ctx, value);
+        let value = coerce_indexed_array_set_value(ctx, &property_ty, value, Some(span));
+        ctx.emit_void(
+            Op::ArraySet,
+            vec![property_value.value, index.value, value.value],
+            None,
+            Op::ArraySet.default_effects(),
+            Some(span),
+        );
+        release_property_array_insert_value_after_retain(ctx, &property_ty, value, span);
+        ctx.emit_void(
+            Op::PropSet,
+            vec![object.value, property_value.value],
+            Some(Immediate::Data(data)),
+            Op::PropSet.default_effects(),
+            Some(span),
+        );
+        release_rewritten_property_value_after_retaining_store(
+            ctx,
+            &property_ty,
+            property_value,
+            span,
+        );
+        return;
+    }
+    if let Some(property_ty) =
+        object_property_type(ctx, object.value, property).filter(is_assoc_array_type)
+    {
+        let data = ctx.intern_string(property);
+        let property_value = ctx.emit_value(
+            Op::PropGet,
+            vec![object.value],
+            Some(Immediate::Data(data)),
+            property_ty.clone(),
+            Op::PropGet.default_effects(),
+            Some(span),
+        );
+        let property_value =
+            crate::ir_lower::ownership::acquire_if_refcounted(ctx, property_value, Some(span));
+        let index = lower_expr(ctx, index);
+        let value = lower_expr(ctx, value);
+        ctx.emit_void(
+            Op::HashSet,
+            vec![property_value.value, index.value, value.value],
+            None,
+            Op::HashSet.default_effects(),
+            Some(span),
+        );
+        release_property_array_insert_value_after_retain(ctx, &property_ty, value, span);
+        ctx.emit_void(
+            Op::PropSet,
+            vec![object.value, property_value.value],
+            Some(Immediate::Data(data)),
+            Op::PropSet.default_effects(),
+            Some(span),
+        );
+        release_rewritten_property_value_after_retaining_store(
+            ctx,
+            &property_ty,
+            property_value,
+            span,
+        );
+        return;
+    }
+
+    if let Some(property_ty) = object_property_type(ctx, object.value, property)
+        .filter(|ty| type_satisfies_array_access_for_ir(ctx, ty))
+    {
+        let data = ctx.intern_string(property);
+        let property_value = ctx.emit_value(
+            Op::PropGet,
+            vec![object.value],
+            Some(Immediate::Data(data)),
+            property_ty,
+            Op::PropGet.default_effects(),
+            Some(span),
+        );
+        let index = lower_expr(ctx, index);
+        let value = lower_expr(ctx, value);
+        ctx.emit_void(
+            Op::RuntimeCall,
+            vec![property_value.value, index.value, value.value],
+            None,
+            effects_lookup::runtime_effects(),
+            Some(span),
+        );
+        return;
+    }
+
+    let index = lower_expr(ctx, index);
+    let value = lower_expr(ctx, value);
+    let data = ctx.intern_string(property);
+    ctx.emit_void(
+        Op::RuntimeCall,
+        vec![object.value, index.value, value.value],
+        Some(Immediate::Data(data)),
+        effects_lookup::runtime_effects(),
+        Some(span),
+    );
+}
+
+/// Releases a temporary assigned into an object property after `PropSet` retains or boxes it.
+pub(super) fn release_property_assignment_source_after_retaining_store(
+    ctx: &mut LoweringContext<'_, '_>,
+    property_ty: &PhpType,
+    value: LoweredValue,
+    span: Span,
+) {
+    if !ctx.value_is_owning_temporary(value) {
+        return;
+    }
+    if !property_store_keeps_independent_ref(property_ty, &ctx.builder.value_php_type(value.value))
+    {
+        return;
+    }
+    crate::ir_lower::ownership::release_if_owned(ctx, value, Some(span));
+}
+
+/// Releases an element temporary after a property-array write retains it for storage.
+pub(super) fn release_property_array_insert_value_after_retain(
+    ctx: &mut LoweringContext<'_, '_>,
+    property_ty: &PhpType,
+    value: LoweredValue,
+    span: Span,
+) {
+    let Some(elem_ty) = indexed_property_array_element_type(property_ty) else {
+        return;
+    };
+    if matches!(elem_ty.codegen_repr(), PhpType::Mixed | PhpType::Callable) {
+        return;
+    }
+    if ctx.value_is_owning_temporary(value) {
+        crate::ir_lower::ownership::release_if_owned(ctx, value, Some(span));
+    }
+}
+
+/// Releases the loaded property value after rewriting it through a retaining `PropSet`.
+pub(super) fn release_rewritten_property_value_after_retaining_store(
+    ctx: &mut LoweringContext<'_, '_>,
+    property_ty: &PhpType,
+    property_value: LoweredValue,
+    span: Span,
+) {
+    if property_ty.codegen_repr().is_refcounted() {
+        crate::ir_lower::ownership::release_if_owned(ctx, property_value, Some(span));
+    }
+}
+
+/// Returns whether a property store creates a distinct retained/boxed owner for the value.
+pub(super) fn property_store_keeps_independent_ref(property_ty: &PhpType, value_ty: &PhpType) -> bool {
+    let property_ty = property_ty.codegen_repr();
+    let value_ty = value_ty.codegen_repr();
+    if matches!((&property_ty, &value_ty), (PhpType::Mixed, PhpType::Mixed)) {
+        return false;
+    }
+    if matches!(value_ty, PhpType::Mixed | PhpType::Union(_))
+        && matches!(property_ty, PhpType::Int | PhpType::Bool | PhpType::Float)
+    {
+        return true;
+    }
+    if matches!(property_ty, PhpType::Str) {
+        return true;
+    }
+    property_ty.is_refcounted()
+}
+
+/// Returns the element type for property arrays that use retaining indexed/hash helpers.
+pub(super) fn indexed_property_array_element_type(property_ty: &PhpType) -> Option<PhpType> {
+    match property_ty.codegen_repr() {
+        PhpType::Array(elem_ty) => Some(elem_ty.codegen_repr()),
+        PhpType::AssocArray { value, .. } => Some(value.codegen_repr()),
+        _ => None,
+    }
+}
+
