+++ src/ir_lower/stmt/statement_basics.rs
@@ -1,0 +1,64 @@
+//! Purpose:
+//! Expression-statement cleanup, block lowering, and echo emission.
+//!
+//! Called from:
+//! - `crate::ir_lower::stmt`.
+//!
+//! Key details:
+//! - Preserves statement ordering, CFG shape, EIR effects, and ownership contracts.
+
+use super::*;
+
+/// Releases a discarded expression-statement result when it may own temporary storage.
+pub(super) fn release_expr_statement_result(
+    ctx: &mut LoweringContext<'_, '_>,
+    value: LoweredValue,
+    span: Span,
+) {
+    if ctx.value_is_owning_temporary(value) {
+        crate::ir_lower::ownership::release_if_owned(ctx, value, Some(span));
+    }
+}
+
+/// Emits the statement-boundary concat-buffer reset expected by the ASM backend.
+pub(super) fn lower_statement_concat_reset(ctx: &mut LoweringContext<'_, '_>, span: Span) {
+    if span.line == 0 {
+        return;
+    }
+    ctx.emit_void(
+        Op::ConcatReset,
+        vec![],
+        None,
+        Op::ConcatReset.default_effects(),
+        Some(span),
+    );
+}
+
+/// Lowers a sequence of statements until the current block terminates.
+pub(super) fn lower_block(ctx: &mut LoweringContext<'_, '_>, body: &[Stmt]) {
+    for stmt in body {
+        lower_stmt(ctx, stmt);
+        if ctx.builder.insertion_block_is_terminated() {
+            break;
+        }
+    }
+}
+
+/// Emits EIR for `echo`.
+pub(super) fn lower_echo(ctx: &mut LoweringContext<'_, '_>, expr: &Expr, span: Span) {
+    let value = lower_expr(ctx, expr);
+    if ctx.builder.insertion_block_is_terminated() {
+        return;
+    }
+    ctx.emit_void(
+        Op::EchoValue,
+        vec![value.value],
+        None,
+        Op::EchoValue.default_effects(),
+        Some(span),
+    );
+    if ctx.value_is_owning_temporary(value) {
+        crate::ir_lower::ownership::release_if_owned(ctx, value, Some(span));
+    }
+}
+
