+++ src/ir_lower/stmt/typed_foreach.rs
@@ -1,0 +1,272 @@
+//! Purpose:
+//! Typed assignments and foreach iterator lowering.
+//!
+//! Called from:
+//! - `crate::ir_lower::stmt`.
+//!
+//! Key details:
+//! - Preserves statement ordering, CFG shape, EIR effects, and ownership contracts.
+
+use super::*;
+
+/// Lowers an assignment with a declared type.
+pub(super) fn lower_typed_assign(
+    ctx: &mut LoweringContext<'_, '_>,
+    type_expr: &crate::parser::ast::TypeExpr,
+    name: &str,
+    value: &Expr,
+    span: Span,
+) {
+    let direct_closure = matches!(value.kind, ExprKind::Closure { .. });
+    ctx.clear_pending_static_callable_result();
+    let php_type = ctx.type_expr_to_php_type_for_value(type_expr);
+    let static_callable = static_callable_binding_for_expr(ctx, value);
+    let reflected_class = reflection_class_binding_for_expr(ctx, value);
+    let reflected_property = reflection_property_binding_for_expr(ctx, value);
+    let fiber_start_sig = crate::ir_lower::fibers::start_sig_for_expr(ctx, value);
+    let callable_array = lower_callable_array_for_assignment(ctx, value, static_callable.as_ref());
+    let lowered = callable_array
+        .as_ref()
+        .map(|assignment| assignment.value)
+        .unwrap_or_else(|| lower_expr(ctx, value));
+    let lowered = coerce_typed_assign_value(ctx, lowered, &php_type, span);
+    ctx.declare_local(name, php_type.clone());
+    ctx.store_local(name, lowered, php_type, Some(span));
+    let callable_result = if direct_closure {
+        ctx.take_pending_static_callable_result()
+    } else {
+        ctx.clear_pending_static_callable_result();
+        None
+    };
+    let static_callable = callable_array
+        .map(|assignment| assignment.target)
+        .or(static_callable)
+        .or(callable_result);
+    if let Some(target) = static_callable {
+        ctx.bind_static_callable_local(name, target);
+    }
+    if let Some(reflected_class) = reflected_class {
+        ctx.bind_reflection_class_local(name, reflected_class);
+    }
+    if let Some((reflected_class, reflected_property)) = reflected_property {
+        ctx.bind_reflection_property_local(name, reflected_class, reflected_property);
+    }
+    if let Some(sig) = fiber_start_sig {
+        ctx.bind_fiber_start_sig(name, sig);
+    }
+}
+
+/// Coerces a typed local assignment into the storage shape required by the declared type.
+pub(super) fn coerce_typed_assign_value(
+    ctx: &mut LoweringContext<'_, '_>,
+    value: LoweredValue,
+    php_type: &PhpType,
+    span: Span,
+) -> LoweredValue {
+    let target_ty = php_type.codegen_repr();
+    let source_ty = ctx.builder.value_php_type(value.value).codegen_repr();
+    if source_ty == target_ty {
+        return value;
+    }
+    match target_ty {
+        PhpType::Mixed => ctx.box_value_as_mixed(value, PhpType::Mixed, Some(span)),
+        _ => value,
+    }
+}
+
+/// Lowers a `foreach` loop using high-level iterator opcodes.
+pub(super) fn lower_foreach(
+    ctx: &mut LoweringContext<'_, '_>,
+    array: &Expr,
+    key_var: Option<&str>,
+    value_var: &str,
+    value_by_ref: bool,
+    body: &[Stmt],
+    loop_span: Span,
+) {
+    // Apply the checker-computed loop header contract before lowering the source expression so
+    // an iterated-and-mutated array is loaded with its stable payload representation.
+    apply_loop_storage_contracts(ctx, loop_span, Some(array.span));
+    let source = lower_expr(ctx, array);
+    let source_php_ty = ctx.builder.value_php_type(source.value);
+    let source_ty = source_php_ty.codegen_repr();
+    let key_needs_null_init = key_var.is_some_and(|name| !ctx.local_slots.contains_key(name));
+    let value_needs_null_init = !ctx.local_slots.contains_key(value_var);
+    // A foreach over a concretely-indexed array (`Array` of a non-Mixed element
+    // type) always yields integer keys, even though `Op::IterCurrentKey` lowers
+    // the key as Mixed. Tag the key local so a `$dst[$key] = ...` write coerces
+    // the int-valued Mixed key to int instead of promoting the destination to a
+    // hash. Generic `Array(Mixed)`, `AssocArray`, `Mixed`, and `Union` sources
+    // may carry string keys and are left untagged so the write promotes.
+    if let Some(key_var) = key_var {
+        if let PhpType::Array(elem_ty) = &source_php_ty {
+            if !matches!(elem_ty.as_ref(), PhpType::Mixed) {
+                ctx.mark_foreach_int_key(key_var);
+            }
+        }
+    }
+    let iterator = ctx.emit_value(
+        Op::IterStart,
+        vec![source.value],
+        value_by_ref.then_some(Immediate::Bool(true)),
+        PhpType::Iterable,
+        Op::IterStart.default_effects(),
+        Some(array.span),
+    );
+    if let Some(key_var) = key_var {
+        initialize_foreach_mixed_local_if_needed(ctx, key_var, key_needs_null_init, array.span);
+    }
+    if value_by_ref {
+        let value_ty = foreach_ref_value_type(&source_ty);
+        ctx.declare_local(value_var, value_ty.clone());
+        ctx.set_local_type(value_var, value_ty);
+        if !value_needs_null_init {
+            ctx.mark_local_initialized(value_var);
+            if !ctx.is_ref_bound_local(value_var) {
+                ctx.promote_local_ref_cell(value_var, Some(array.span));
+            }
+        }
+    } else {
+        let value_ty = foreach_value_type(&source_ty);
+        if value_ty == PhpType::Mixed {
+            initialize_foreach_mixed_local_if_needed(
+                ctx,
+                value_var,
+                value_needs_null_init,
+                array.span,
+            );
+        } else if value_needs_null_init {
+            ctx.declare_local(value_var, value_ty.clone());
+            ctx.set_local_type(value_var, value_ty);
+        }
+    }
+    let header = ctx.builder.create_named_block("foreach.next", Vec::new());
+    let body_block = ctx.builder.create_named_block("foreach.body", Vec::new());
+    let exit = ctx.builder.create_named_block("foreach.exit", Vec::new());
+    branch_to(ctx, header);
+
+    ctx.builder.position_at_end(header);
+    let has_next = ctx.emit_value(
+        Op::IterNext,
+        vec![iterator.value],
+        None,
+        PhpType::Bool,
+        Op::IterNext.default_effects(),
+        Some(array.span),
+    );
+    ctx.builder.terminate(Terminator::CondBr {
+        cond: has_next.value,
+        then_target: body_block,
+        then_args: Vec::new(),
+        else_target: exit,
+        else_args: Vec::new(),
+    });
+
+    ctx.clear_static_callable_locals();
+    ctx.builder.position_at_end(body_block);
+    let cleanup = ctx
+        .value_is_owning_temporary(source)
+        .then_some(LoopCleanup {
+            value: source,
+            span: array.span,
+        });
+    ctx.loop_stack.push(LoopFrame {
+        break_block: exit,
+        continue_block: header,
+        cleanup,
+    });
+    if let Some(key_var) = key_var {
+        let key = ctx.emit_value(
+            Op::IterCurrentKey,
+            vec![iterator.value],
+            None,
+            PhpType::Mixed,
+            Op::IterCurrentKey.default_effects(),
+            Some(array.span),
+        );
+        ctx.store_local(key_var, key, PhpType::Mixed, Some(array.span));
+    }
+    if value_by_ref {
+        let slot = ctx.declare_local(value_var, foreach_ref_value_type(&source_ty));
+        ctx.release_ref_cell_owner(value_var, Some(array.span));
+        ctx.emit_void(
+            Op::IterCurrentValueRef,
+            vec![iterator.value],
+            Some(Immediate::LocalSlot(slot)),
+            Op::IterCurrentValueRef.default_effects(),
+            Some(array.span),
+        );
+        ctx.mark_ref_bound_local(value_var);
+        ctx.mark_local_initialized(value_var);
+    } else {
+        let value_ty = foreach_value_type(&source_ty);
+        let value = ctx.emit_value(
+            Op::IterCurrentValue,
+            vec![iterator.value],
+            None,
+            value_ty.clone(),
+            Op::IterCurrentValue.default_effects(),
+            Some(array.span),
+        );
+        ctx.store_local(value_var, value, value_ty, Some(array.span));
+    }
+    lower_block(ctx, body);
+    ctx.loop_stack.pop();
+    branch_to(ctx, header);
+    ctx.builder.position_at_end(exit);
+    ctx.clear_static_callable_locals();
+    // Release the source when it is a fresh owning temporary (e.g. `foreach
+    // (explode(...) as $p)` or a literal array): the iterator borrows it for the
+    // duration of the loop, so nothing else frees it once iteration ends. (For an
+    // array the iterator aliases the source, so it must NOT be released separately
+    // — that would double-free.)
+    if ctx.value_is_owning_temporary(source) {
+        crate::ir_lower::ownership::release_if_owned(ctx, source, Some(array.span));
+    }
+}
+
+/// Returns the by-value foreach local type when Phase 04 can keep a concrete element.
+pub(super) fn foreach_value_type(source_ty: &PhpType) -> PhpType {
+    match source_ty.codegen_repr() {
+        PhpType::Array(elem) => match elem.codegen_repr() {
+            PhpType::Callable => PhpType::Callable,
+            PhpType::Object(class_name) => PhpType::Object(class_name),
+            elem @ (PhpType::Int | PhpType::Float | PhpType::Str | PhpType::Bool) => elem,
+            _ => PhpType::Mixed,
+        },
+        PhpType::Object(class_name) if class_name == "Phar" || class_name == "PharData" => {
+            PhpType::Object("PharFileInfo".to_string())
+        }
+        _ => PhpType::Mixed,
+    }
+}
+
+/// Returns the local value type used when a foreach binds the value by reference.
+pub(super) fn foreach_ref_value_type(source_ty: &PhpType) -> PhpType {
+    match source_ty.codegen_repr() {
+        PhpType::Array(elem) => *elem,
+        PhpType::AssocArray { value, .. } => *value,
+        _ => PhpType::Mixed,
+    }
+}
+
+/// Initializes a fresh foreach loop variable to boxed null before the first iteration.
+pub(super) fn initialize_foreach_mixed_local_if_needed(
+    ctx: &mut LoweringContext<'_, '_>,
+    name: &str,
+    needs_init: bool,
+    span: Span,
+) {
+    if !needs_init {
+        return;
+    }
+    // This setup can run once per outer-loop iteration at runtime, overwriting
+    // the loop variable. `store_local` owns the carried release: it frees the
+    // previous runtime occupant when this synthetic store is loop-carried.
+    ctx.declare_local(name, PhpType::Mixed);
+    ctx.set_local_type(name, PhpType::Mixed);
+    let null = emit_null_value(ctx, Some(span));
+    let boxed = ctx.box_value_as_mixed(null, PhpType::Mixed, Some(span));
+    ctx.store_foreach_initializer_local_only(name, boxed, PhpType::Mixed, Some(span));
+}
+
